下までスクロール

+++ src/main/java/com/example/breakserver/service/DocumentSentiment.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class DocumentSentiment{
-    int magnitude;
-    int score;
+    Double magnitude;
+    Double score;
 }
